Refactor account selection
When rh-aws-saml-login is used programmatically, it is the caller's
responsibility to provide an account name (and role, optionally). On the
other hand, the caller does not want to face the risk of an interactive
CLI utility (fzf) being spawned.

This commit splits selecting an account for a known account name (and
role) from the task of determining the account name (and role) when
rh-aws-saml-login is used as a CLI tool.

It also moves the responsibility of handling special `account_name`
values (i.e. `None` and `"."`) into a single function
(`get_aws_account`). The responsibility was originally split between
the original `select_aws_account` and `main` functions.