Add 'uv' to 'dev' dependency group
The project uses `Makefile` targets to define linters and tests to run.
The targets are implemented using `uv`. Adding `uv` as a dev dependency
makes sure that developers will be able to use the `Makefile` targets
when they install the `dev` dependency group to their development
environment, regardless of what tools they use to create the development
environment.